refactor(cli): extract upgrade into api/upgrade (thin handler + injected logger)
Move the full version-to-version upgrade pipeline (codemods + agent-docs
refresh) out of the CLI handler into api/upgrade/upgrade.mjs. It performs the
effect and returns a {type,data} receipt (upgrade.list | upgrade.status |
upgrade.run), throwing AstryxError on failure.

Human progress is emitted through an injectable logger (lib/cli-logger.mjs):
the CLI passes a term-log-backed logger, programmatic callers get the silent
noopLogger default — so a scripted upgrade() never spams the console while the
CLI keeps byte-identical output.

The CLI handler is now a thin wrapper (861 -> 67 lines): parse flags, wire the
logger, render result / exit code. Only AstryxError is intercepted (routed to
jsonError / exit code); unexpected throws still reach the top-level boundary.

- lib/cli-logger.mjs: noopLogger + termLogger
- api/index.mjs, types/api.d.ts: export upgrade + UpgradeOptions
- parity: promote upgrade (--list + arg-validation errors) to full api<->cli
  parity (no longer CLI-only)

Verified: upgrade unit suites (23), typecheck:strict + json-api, eslint,
json-smoke, api<->cli parity 332/0 (upgrade byte-identical).
